the logic is in place for the delete button so users can only delete what they have posted

+++ src/scripts/messages/messageEvents.js
@@ -6,7 +6,11 @@
 import messagesComponent from "./messagesComponent.js"
 import messagesAPI from "./messagesData.js"
 import messageList from "./messagesList.js"
+
+const userId = 5;
+const userName = "Gavin Swofford"
 const messageEventListener = () => {
+    
     const messageEventLocation = document.querySelector(".messages")
     const messageInputLocation = document.querySelector(".messageInputSection")
 
@@ -20,7 +24,9 @@ const messageEventListener = () => {
         //must have the hidden value not be on the page in order to not have conflicts
         else if(clickEvent.target.className == "message--SubmitButton"  && !document.querySelector("#hiddenIdValue")) {
             const inputLocation = document.querySelector(".message--Input") 
-            const messageInputValue = {"message": inputLocation.value}     
+            const messageInputValue = {"message": inputLocation.value, "userID": userId, "userName": userName} 
+            //=========================================================================================================================
+            //WILL NEED TO IMPORT A USER ID TO MATCH AND SAVE TO EACH ENTRY    
                  console.log(messageInputValue)
             messagesAPI.messagePostData(messageInputValue)
             .then(() =>{
@@ -34,15 +40,19 @@ const messageEventListener = () => {
 
         //delete message event
         else if(clickEvent.target.id.startsWith("message--DeleteButton--")) {
-            const cardDelete = event.target.id.split("--DeleteButton--")[1];
+            const cardDeleteIdAndUserId = event.target.id.split("--DeleteButton--")[1];
+            const cardDelete = cardDeleteIdAndUserId.split("--")[0]
+            const cardUserID = cardDeleteIdAndUserId.split("--")[1]
+            console.log(cardUserID, "this is the user ID!!!")
             console.log(cardDelete)
-
-            messagesAPI.deletePostData(cardDelete).then(() =>{
-                messagesAPI.messagesGetData().then(() => {
-                    messageList()
+            if(userId == cardUserID) {
+             messagesAPI.deletePostData(cardDelete).then(() =>{
+                 messagesAPI.messagesGetData().then(() => {
+                     messageList()
                     
-                })
-            })
+                 })
+             })
+            }
         }
         //populating the input boxes at the message
         else if(clickEvent.target.id.startsWith("message--EditButton--") && !document.querySelector("#hiddenIdValue")) {
@@ -58,7 +68,7 @@ const messageEventListener = () => {
             const editCheck = document.querySelector(".message--Input")
             if(editCheck.value != "") {
                 const messageObjectID = document.querySelector("#hiddenIdValue").value
-                messagesAPI.updateEditMessage(messageObjectID, {"message": editCheck.value})
+                messagesAPI.updateEditMessage(messageObjectID, {"message": editCheck.value, "userID": userId, "userName": userName})
                 .then(() => {
                     messagesAPI.messagesGetData().then(() => {
                         messageList()
